refactor engine classes
* Combine token classes
* Combine instance classes to remove distinction between “exercise” and
“runnable code”
